MainMenu: Add Continue and Restart buttons
In the game state, add a can_restore() method that checks if the config
file has sections. If so, it can be continued. Also add a clear() method
to empty the config file.

Add a Continue button to the main menu. Show it only if the game can be
continued. And change Start to "Restart" in that case.

When (re)starting, clear the config file.

Focus the Continue button if it is visible. Otherwise focus the Start
button as before.

Also: Add 2 exceptions to persisting the current scene: the title and
the intro.